fix status logged out

+++ src/components/ProfilIcon.tsx
@@ -7,7 +7,11 @@ import {
     DropdownTrigger,
 } from '@nextui-org/react';
 import { useRouter } from 'next/router';
+import { useDispatch } from 'react-redux';
+import { useSelector } from 'react-redux';
 import useAuth from 'src/hooks/useAuth';
+import { setShow } from 'src/redux/slice/cardShow';
+import ModalNotif from './Modals/ModalNotif';
 
 export default function ProfileIcon({
     picture = 'https://ps.w.org/user-avatar-reloaded/assets/icon-256x256.png?rev=2540745',
@@ -17,15 +21,30 @@ export default function ProfileIcon({
 }: any) {
     const { onLogout } = useAuth();
     const route = useRouter();
+    const isShowNotif = useSelector((store: any) => store.show.show);
+    const dispatch = useDispatch();
     const handleLogout = () => {
-        onLogout('logout', () => {
-            route.push('/Auth/login');
-            localStorage.clear();
+        onLogout('logout', (res: any) => {
+            if (res.status === 200) {
+                dispatch(setShow());
+                setTimeout(() => {
+                    route.push('/Auth/login');
+                    localStorage.clear();
+                }, 2000);
+            }
         });
     };
 
     return (
         <>
+            {isShowNotif && (
+                <ModalNotif
+                    modal={{
+                        head: 'Logout Success',
+                        text: 'You have been logged out, goodbye!',
+                    }}
+                />
+            )}
             <Dropdown
                 showArrow
                 radius="sm"
